# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Bridge API
  version: 1.125.0
  description: 'Bridge — federated, self-hosted communication platform REST API.


    **Authentication:** Tüm korumalı endpoint''ler `Authorization: Bearer <accessToken>` header''ı gerektirir.

    Access token''lar kısa ömürlüdür (15 dk). `POST /api/refresh` ile yenilenebilir (httpOnly cookie).


    **Rate Limiting:** Endpoint başına granüler rate limit uygulanır. 429 döndüğünde `Retry-After` header''ını kontrol edin.


    **Federation:** Bridge sunucular arası ActivityPub protokolü kullanır. Federation endpoint''leri `/api/federation` altındadır.

    '
  contact:
    name: Bridge Contributors
    url: https://github.com/bridge-app/bridge
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: https://your-bridge-instance.com/api
  description: Production
- url: http://localhost:3001/api
  description: Local development
...
paths:
...
  /servers/{sid}/boosts:
    get:
      tags:
      - Boosts
      summary: Sunucu boost listesi
      security:
      - bearerAuth: []
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Boost listesi
      operationId: get_servers_sid_boosts
    post:
      tags:
      - Boosts
      summary: Sunucuyu boost et
      security:
      - bearerAuth: []
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Boost başarılı
        409:
          description: Zaten boost edilmiş
      operationId: post_servers_sid_boosts
    delete:
      tags:
      - Boosts
      summary: Boost geri al
      security:
      - bearerAuth: []
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Boost kaldırıldı
      operationId: delete_servers_sid_boosts
  /servers/vanity/{slug}:
    get:
      tags:
      - Boosts
      summary: Vanity URL ile sunucu bul
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Sunucu bilgisi
        404:
          description: Vanity URL bulunamadı
      operationId: get_servers_vanity_slug
  /servers/{sid}/vanity:
    patch:
      tags:
      - Boosts
      summary: Vanity URL güncelle
      security:
      - bearerAuth: []
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - vanityUrl
              properties:
                vanityUrl:
                  type:
                  - string
                  - 'null'
                  minLength: 3
                  maxLength: 32
                  pattern: ^[a-z0-9-]+$
      responses:
        200:
          description: Vanity URL güncellendi
      operationId: patch_servers_sid_vanity
...
  /servers/{sid}/emojis:
    get:
      tags:
      - Servers
      summary: Ozel emojileri sayfalı listele
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: sid
        required: true
        schema:
          type: string
      - in: query
        name: limit
        schema:
          type: integer
          default: 50
      responses:
        200:
          description: Sayfalı emoji listesi
      operationId: get_servers_sid_emojis
  /servers/{sid}/emojis/{emojiId}:
    patch:
      tags:
      - Servers
      summary: Emoji adini guncelle
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: sid
        required: true
        schema:
          type: string
      - in: path
        name: emojiId
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
      responses:
        200:
          description: Guncellendi
        403:
          $ref: '#/components/responses/Forbidden'
      operationId: patch_servers_sid_emojis_emojiId
    delete:
      tags:
      - Servers
      summary: Emojiyi sil
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: sid
        required: true
        schema:
          type: string
      - in: path
        name: emojiId
        required: true
        schema:
          type: string
      responses:
        200:
          description: Silindi
        403:
          $ref: '#/components/responses/Forbidden'
      operationId: delete_servers_sid_emojis_emojiId
  /servers/{sid}/emojis/all:
    get:
      tags:
      - Servers
      summary: Sunucu tum ozel emojilerini getir (sayfalama yok)
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: sid
        required: true
        schema:
          type: string
      responses:
        200:
          description: Tum emoji listesi
      operationId: get_servers_sid_emojis_all
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. POST /api/login ile alınır, süresi 15 dakikadır.
    botToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bot token: `Authorization: Bot brg_bot_…` (sunucu sahibi bot oluştururken bir kez gösterilir).'
...
  responses:
...
    Forbidden:
      description: Yetersiz izin
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
security:
- bearerAuth: []